refactor(api): moved json utility in internal package
